integrate Systems.h, Drones.h and Tasks.h
