+++ CustomCrew.cpp
@@ -9,7 +9,6 @@
 #include "ShipUnlocks.h"
 #include "CustomEvents.h"
 #include "CustomSystems.h"
-#include "Tasks.h"
 
 #include <boost/lexical_cast.hpp>
 #include <boost/algorithm/string.hpp>
@@ -5369,7 +5368,7 @@ HOOK_METHOD(CrewAI, PrioritizeTask, (CrewTask task, int crewId) -> int)
 {
     LOG_HOOK("HOOK_METHOD -> CrewAI::PrioritizeTask -> Begin (CustomCrew.cpp)\n")
     if (crewId == -1) return super(task, crewId);
-    
+
     CrewMember* crew = crewList[crewId];
     if (task.taskId == TASK_MANNING && !crew->CanMan())
     {
@@ -6825,7 +6824,7 @@ HOOK_METHOD_PRIORITY(ShipManager, CountPlayerCrew, 9999, () -> int)
     LOG_HOOK("HOOK_METHOD_PRIORITY -> ShipManager::CountPlayerCrew -> Begin (CustomCrew.cpp)\n")
     int ret = 0;
     for (auto& crew: vCrewList)
-    {   
+    {
         auto ex = CM_EX(crew);
 
         bool canTeleport;
@@ -6843,7 +6842,7 @@ HOOK_METHOD_PRIORITY(ShipManager, CountPlayerCrew, 9999, () -> int)
 HOOK_METHOD(CrewMember, RestorePosition, () -> bool)
 {
     LOG_HOOK("HOOK_METHOD -> CrewMember::RestorePosition -> Begin (CustomCrew.cpp)\n")
-    
+
     Slot station;
     if (!bDead && savedPosition.roomId != -1 && (station = FindSlot(savedPosition.roomId, savedPosition.slotId, false), station.roomId > -1) && station.slotId > -1)
     {
@@ -6911,7 +6910,7 @@ HOOK_METHOD(CrewMember, RestorePosition, () -> bool)
             }
         }
     }
-    
+
     return false;
 }
 
@@ -6947,7 +6946,7 @@ HOOK_METHOD(ShipGraph, FindPath, (Point p1, Point p2, int shipId) -> Path)
             {
                 int slot = shipManager->ship.vRoomList[backupRoom.first]->GetEmptySlot(false);
                 ret = super(p1, GetSlotWorldPosition(slot, backupRoom.first), shipId);
-                if (ret.distance != -1.0) 
+                if (ret.distance != -1.0)
                 {
                     g_partitionDestRoomId = backupRoom.first;
                     g_partitionDestSlotId = slot;
