--- Drones.h
@@ -1,14 +1,0 @@
-#pragma once
-enum DroneType
-{
-    
-    DRONE_DEFENSE = 0,
-    DRONE_COMBAT = 1,
-    DRONE_REPAIR = 2,
-    DRONE_BATTLE = 3,
-    DRONE_BOARDER = 4,
-    DRONE_SHIP_REPAIR = 5,
-    DRONE_FIGHTER_TEST = 6,
-    DRONE_SHIELD = 7,
-    DRONE_TOTAL = 8
-};
