+++ Global.h
@@ -6,7 +6,7 @@
 #include "zhl.h"
 #include "FTLGame.h"
 
-#include "Systems.h"
+#include "Enums.h"
 #include "Colors.h"
