--- Systems.h
@@ -1,29 +1,0 @@
-#pragma once
-enum SystemId
-{
-    SYS_INVALID = -1,
-    SYS_SHIELDS = 0,    
-    SYS_ENGINES = 1,    
-    SYS_OXYGEN = 2,     
-    SYS_WEAPONS = 3,    
-    SYS_DRONES = 4,     
-    SYS_MEDBAY = 5,     
-    SYS_PILOT = 6,      
-    SYS_SENSORS = 7,    
-    SYS_DOORS = 8,      
-    SYS_TELEPORTER = 9, 
-    SYS_CLOAKING = 10,   
-    SYS_ARTILLERY = 11,  
-    SYS_BATTERY = 12,    
-    SYS_CLONEBAY = 13,   
-    SYS_MIND = 14,       
-    SYS_HACKING = 15,    
-    
-    SYS_ALL = 16,
-    SYS_REACTOR = 17,
-    SYS_RANDOM = 18,
-    SYS_ROOM = 19,
-
-    SYS_TEMPORAL = 20,
-    SYS_CUSTOM_FIRST = 21,
-};
