--- Tasks.h
@@ -1,15 +1,0 @@
-#pragma once
-enum TaskId
-{
-    TASK_MANNING = 0,
-    TASK_REPAIRING = 1,
-    TASK_FIRE = 2,
-    TASK_BREACH = 3,
-    TASK_NOTHING = 4,
-    TASK_DEAD = 5,
-    TASK_HEAL = 6,
-    TASK_BOARDERS = 7,
-    TASK_TELEPORTING = 8,
-    TASK_KILLING = 9,
-    TASK_SABOTAGING = 10
-};
